開発と本番環境で処理変更

+++ src/app/Http/Controllers/AdminController.php
@@ -106,14 +106,16 @@ class AdminController extends Controller
 
         $file_extension = $request->file('shop_image')->getClientOriginalExtension();
         $file = $request->file('shop_image');
-        $dir = 'rese_restaurant/';
-        // $upload_file = Storage::disk('s3')->putFileAs($dir, $file, 'restaurant_'.$restaurant_id.'.'.$file_extension);
-        // $restaurant->update(['storage_image' => $upload_file]);
-        // dd($upload_file);
-        $request->file('shop_image')->storeAs('public/restaurant_images', 'restaurant_'.$restaurant_id.'.'.$file_extension);
+        $dir = 'rese_restaurant';
 
-        $restaurant->update(['storage_image' => 'storage/restaurant_images/restaurant_'.$restaurant_id.'.'.$file_extension]);
-
+        if(config('app.env') == 'local')
+        {
+            $request->file('shop_image')->storeAs('public/restaurant_images', 'restaurant_'.$restaurant_id.'.'.$file_extension);
+            $restaurant->update(['storage_image' => 'storage/restaurant_images/restaurant_'.$restaurant_id.'.'.$file_extension]);
+        }else{
+            $upload_file = Storage::disk('s3')->putFileAs($dir, $file, 'restaurant_'.$restaurant_id.'.'.$file_extension);
+            $restaurant->update(['storage_image' => $upload_file]);
+        }
 
         return redirect('/restaurant_owner');
     }
@@ -121,36 +123,22 @@ class AdminController extends Controller
     public function shop_update(Request $request)
     {
         $restaurant_owner = RestaurantOwner::with('restaurant')->find(Auth::guard('restaurant_owners')->id());
-        $storage_image = $restaurant_owner->restaurant->storage_image;
-        $path = substr($storage_image, 7, 50);
-        Storage::disk('public')->delete($path);
-        // dd($path);
-        // dd('u');
-        // dd(Storage::disk('public'));
-        // dd($storage_image);
-        // dd(Storage::url($storage_image));
-        // Storage::disk('public')->delete('/restaurant_images'.'/'.'restaurant_68.jpg');
-        // $file = Storage::allFiles('public/');
-        // dd($file);
-        // dd(Storage::disk('public')->exists('/restaurant_images'.'/'.'restaurant_32.jpeg'));
-        // dd('i');
-        // dd(Storage::disk('public'));
-        // dd($storage_image);
-        // $file = $request->file('shop_image');
         $file_extension = $request->file('shop_image')->getClientOriginalExtension();
-        // $dir = 'rese_restaurant/';
-        // $storage_image->delete();
-        // dd('K');
-        $request->file('shop_image')->storeAs('public/restaurant_images', 'restaurant_'.$restaurant_owner->restaurant_id.'.'.$file_extension);
-        // $request->file('shop_image')->storeAs('public/'.$dir, $file_name);
+
+        if(config('app.env') == 'local')
+        {
+            $storage_image = $restaurant_owner->restaurant->storage_image;
+            $path = substr($storage_image, 7, 50);
+            Storage::disk('public')->delete($path);
+            $request->file('shop_image')->storeAs('public/restaurant_images', 'restaurant_'.$restaurant_owner->restaurant_id.'.'.$file_extension);
+            Restaurant::find($restaurant_owner->restaurant_id)->update([
+                'name' => $request->shop_name,
+                'name_of_reading_kana' => $request->name_of_reading_kana,
+                'storage_image' => 'storage/restaurant_images/restaurant_'.$restaurant_owner->restaurant_id.'.'.$file_extension]);
+        }else{
+
+        }
         $restaurant_owner->update(['name' => $request->shop_name.'代表者']);
-        Restaurant::find($restaurant_owner->restaurant_id)->update([
-            'name' => $request->shop_name,
-            'name_of_reading_kana' => $request->name_of_reading_kana,
-            'storage_image' => 'storage/restaurant_images/restaurant_'.$restaurant_owner->restaurant_id.'.'.$file_extension]);
-
-
-        // $restaurant = Restaurant::find($restaurant_owner->restaurant_id)->update(['name' => $request->shop_name, 'name_of_reading_kana' => $request->name_of_reading_kana, 'storage_image' => 'storage/'.$dir.'/'.$file_name]);
 
         return redirect('/restaurant_owner');
     }
